modularized movement functions to reduce amount of code written in k.scene()
